refactor(reef): use shared scroll area imports
